@@ -1,9 +1,15 @@
 {
-  "name":    "node-uptime",
-  "description": "Remote monitoring for HTTP applications",
+  "name":    "hoot",
+  "description": "Remote monitoring for ISD HTTP applications",
   "version": "3.2.0",
-  "author": "Francois Zaninotto",
+  "author": "Ashley Narcisse-Austin",
+  "private": true,
   "dependencies": {
+    "babel-core": "^6.11.4",
+    "babel-eslint": "^6.1.2",
+    "babel-plugin-add-module-exports": "^0.2.1",
+    "babel-polyfill": "^6.9.1",
+    "babel-preset-es2015": "^6.9.0",
     "mongoose":  "3.6.7",
     "mongoose-lifecycle": "1.0.0",
     "express":   "3.2.0",
@@ -28,11 +34,11 @@
   "repository": "https://github.com/fzaninotto/uptime",
   "license": "MIT",
   "engines": {
-    "node": "0.10.x",
-    "npm": "1.2.x"
+    "node": "6.3.x",
+    "npm": "3.x.x"
   },
   "scripts": {
     "test": "node_modules/.bin/mocha test/lib/",
-    "start": "node app.js"
+    "start": "nodemon app.js --exec babel-node --presets es2015"
   }
 }
