@@ -235,6 +235,10 @@ Monitor.prototype.applyApiHttpOptions = function(options) {
  * @param {Object} Configuration object
  * @api   public
  */
-exports.createMonitor = function(config) {
-  return new Monitor(config);
-};
+// exports.createMonitor = function(config) {
+//   return new Monitor(config);
+// };
+
+
+export default Monitor;
+// export default config => new Monitor(config);
